Release 0.0.8

+++ src/main/kotlin/DFReaderClockUSec.kt
@@ -1,7 +1,7 @@
-/*
+/**
  * DFReaderClockUSec
  * Copyright (C) 2021 Hitec Commercial Solutions
- * Author, Stephen Woerner
+ * @author Stephen Woerner
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -32,7 +32,7 @@ class DFReaderClockUSec : DFReaderClock() {
     /**
      * work out time basis for the log - even newer style
      */
-    fun findTimeBase(gps: DFMessage, first_us_stamp: Int) {
+    fun findTimeBase(gps: DFMessage, first_us_stamp: Long) {
         val t = gpsTimeToTime(gps.GWk!!, gps.GMS!!)
         timebase = (t - gps.TimeUS!! * 0.000001)
 //         this ensures FMT messages get appropriate timestamp:
